+++ build.gradle
@@ -66,7 +66,7 @@ jsonProviders.forEach {
 dependencies {
     formatImplementation 'com.sanctionco.jmail:jmail:1.6.3'
 
-    testImplementation 'org.eclipse.parsson:parsson:1.1.7'
+    jakartaJsonTestImplementation 'org.eclipse.parsson:parsson:1.1.7'
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.11.4'
     testImplementation 'org.junit.jupiter:junit-jupiter:5.11.4'
     testImplementation 'org.assertj:assertj-core:3.27.0'
